Include invalid tokens in diagnostics
Return a list of errors from the tokenization step, as a second
returned value after the list of tokens, instead of just silently
stopping the tokenization. The error list contains at most one error
for a token mismatch (the tokenization currently stops on the first
error).

This error gets converted to a LuaError in build_lua_doc, so that it
gets included with the diagnostics notifications.

This means that characters that do not match any regexp in tokenize.py
will be shown as errors while editing a lua-file.